Perimeter completed
Final version of perimeter and angle data definitions.
Plan for split is to have function that crawls for flase convex?,
"cycles" list to make the angle the first in perim, look for conecting
point for sub-perim, and split perim into two perims. Recurse this until
perim-convex? is true for all perim